IMPLEMENTED THE SINGLE IMAGE VIEW
A user can now view individual images and view details about location, date posted, tags & categories
Mostly static stuff so far.

I HAVE NOT YET IMPLEMENTED the search functions: That is next in line.

I also want to implement the ranking system to view most popular photos...

ASs mentioned in previous commit, the location field is a basic CharField as of now but this will change when I EITHER implement the google maps OR a registerd view with an assosiated drop down."